Refactoring out hard dependencies
Added an option for specifying a specific queue from the command line
(such as heavy, light, etc.).

Refactored the base Sonido class so that most of its outside
dependencies are injected rather than created in the constructor. Also
made sure that Sonido is typehinted against interfaces, and not
specific classes.

issue 13